Add Formatter namespace
