--- packages/common/src/utils/debounce/debounce.ts
@@ -1,18 +1,0 @@
-export const debounce = <T extends (...args: any[]) => any>(
-  callback: T,
-  wait = 500
-): ((...args: Parameters<T>) => Promise<ReturnType<T>>) => {
-  let timer: NodeJS.Timeout | undefined;
-
-  return (...args: Parameters<T>) => {
-    if (timer) {
-      clearTimeout(timer);
-    }
-    return new Promise<ReturnType<T>>(resolve => {
-      timer = setTimeout(() => {
-        const returnValue = callback(...args) as ReturnType<T>;
-        resolve(returnValue);
-      }, wait);
-    });
-  };
-};
