--- packages/common/src/utils/debounce/index.ts
@@ -1,1 +1,0 @@
-export * from './debounce';
