+++ packages/invoices/src/OutboundShipment/DetailView/SidePanel/PricingSection.tsx
@@ -1,6 +1,6 @@
 import React, { memo } from 'react';
 import {
-  formatTax,
+  Formatter,
   Tooltip,
   Grid,
   DetailPanelSection,
@@ -85,7 +85,7 @@ const ServiceCharges = ({ pricing, isDisabled }: PricingGroupProps) => {
         <PanelField>{c(serviceTotalBeforeTax)}</PanelField>
       </PanelRow>
       <PanelRow sx={{ marginLeft: '10px' }}>
-        <PanelLabel>{`${t('heading.tax')} ${formatTax(tax)}`}</PanelLabel>
+        <PanelLabel>{`${t('heading.tax')} ${Formatter.tax(tax)}`}</PanelLabel>
         <PanelField>
           <TaxEdit
             disabled={!serviceLines?.length || isDisabled}
@@ -135,7 +135,7 @@ const ItemPrices = ({ pricing, isDisabled }: PricingGroupProps) => {
         <PanelField>{c(stockTotalBeforeTax)}</PanelField>
       </PanelRow>
       <PanelRow sx={{ marginLeft: '10px' }}>
-        <PanelLabel>{`${t('heading.tax')} ${formatTax(tax)}`}</PanelLabel>
+        <PanelLabel>{`${t('heading.tax')} ${Formatter.tax(tax)}`}</PanelLabel>
         <PanelField>
           <TaxEdit
             disabled={!outboundLines?.length || isDisabled}
